ECSEL Prototype
Prototype with complete login functionality and fully functioning task
upload, download, and rating forms. Note that the site is in a folder
called "WebForms". I still have to integrate the login UI with the site
, make modifications on the forms for reading log tasks, incorporate
validation controls, and work out a couple minor glitches on the
taskrating page. Tania, please make this look better, it's very
minimalist atm